Move 'slow imports' test to separate file
